pyproject.toml
[tool.poetry]
name = "beda-backend"
version = "0.1.0"
description = ""
authors = ["Beda Team <beda@danwin1210.me>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.63.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.16.19"}
pytest = "^6.2.1"
bcrypt = "^3.2.0"
aerich = "^0.4.3"
aiomysql = "^0.0.21"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.790"
pre-commit = "^2.9.3"
reorder-python-imports = "^2.3.6"
pytest-sugar = "^0.9.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
